# swagger/swagger.yaml
---
swagger: "2.0"
info:
  version: "v1"
  title: "app-service"
tags:
- name: "community api"
  description: "Manage Community"
- name: "sample api"
  description: "Operate hoge"
paths:
  /v1/community:
    post:
      tags:
      - "community api"
      summary: "Post Community"
      description: "insert community"
      operationId: "helloWorld"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/PostCommunityInputDto"
      responses:
        200:
          description: "return status code only"
          schema:
            $ref: "#/definitions/PostCommunityOutputDto"
  /v1/community/list:
    get:
      tags:
      - "community api"
      summary: "GET Communities List"
      description: "return 200"
      operationId: "GetGommunitiesList"
      parameters:
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "return Communities List"
          schema:
            $ref: "#/definitions/GetCommunitiesListOutputDto"
  /v1/sample:
    get:
      tags:
      - "sample api"
      summary: "check"
      description: "return 200"
      operationId: "check"
      parameters: []
      responses:
        200:
          description: "200を返すだけ"
    post:
      tags:
      - "sample api"
      summary: "HelloWorld"
      description: "insert/update into sample"
      operationId: "helloWorld"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/HelloWorldInputDto"
      responses:
        200:
          description: "hello wolrdを返す"
          schema:
            $ref: "#/definitions/HelloWorldOutputDto"
definitions:
  Community:
    type: "object"
    properties:
      communityId:
        type: "integer"
        format: "int32"
      communityName:
        type: "string"
      content:
        type: "string"
      locationId:
        type: "integer"
        format: "int32"
      locationName:
        type: "string"
      note:
        type: "string"
      profileImgUrl:
        type: "string"
      headerImgUrl:
        type: "string"
      tagList:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
      requirement:
        type: "string"
      created:
        type: "string"
        format: "date-time"
      modified:
        type: "string"
        format: "date-time"
  GetCommunitiesListOutputDto:
    type: "object"
    properties:
      communityList:
        type: "array"
        items:
          $ref: "#/definitions/Community"
  HelloWorldInputDto:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
  HelloWorldOutputDto:
    type: "object"
    properties:
      count:
        type: "integer"
        format: "int32"
      message:
        type: "string"
  PostCommunityInputDto:
    type: "object"
    required:
    - "communityName"
    - "locationId"
    properties:
      communityName:
        type: "string"
      content:
        type: "string"
      locationId:
        type: "integer"
        format: "int32"
      note:
        type: "string"
      tagList:
        type: "array"
        items:
          type: "string"
      requirement:
        type: "string"
  PostCommunityOutputDto:
    type: "object"
    properties:
      communityId:
        type: "integer"
        format: "int32"
  Tag:
    type: "object"
    properties:
      tagId:
        type: "integer"
        format: "int32"
      tagName:
        type: "string"
      communityId:
        type: "integer"
        format: "int32"
      tagType:
        type: "integer"
        format: "int32"